test: move twig tests in separate file
